Fix signed  vs unsigned comparisons: strlen
Cast results from strlen to ssize_t as our string handling code uses
pointers whihc are signed. These places should probably replaced by
proper C++ strings at some point.

Related: #3226

+++ lib/header.cc
@@ -1275,7 +1275,7 @@ static int headerMatchLocale(const char *td, const char *l, const char *le)
     const char *fe;
 
     /* First try a complete match. */
-    if (strlen(td) == (le-l) && rstreqn(td, l, (le - l)))
+    if (((ssize_t) strlen(td)) == (le-l) && rstreqn(td, l, (le - l)))
 	return 1;
 
     /* Next, try stripping optional dialect and matching.  */
